+++ src/adaptors/erased.rs
@@ -156,30 +156,103 @@ where
     type Err = Err;
 
     requester_forward! {
-        get_me, log_out, close, get_updates, set_webhook, delete_webhook, get_webhook_info,
-        forward_message, copy_message, send_message, send_photo, send_audio, send_document,
-        send_video, send_animation, send_voice, send_video_note, send_media_group, send_location,
-        edit_message_live_location, edit_message_live_location_inline,
-        stop_message_live_location, stop_message_live_location_inline, send_venue,
-        send_contact, send_poll, send_dice, send_chat_action, get_user_profile_photos,
-        get_file, kick_chat_member, ban_chat_member, unban_chat_member, restrict_chat_member,
-        promote_chat_member, set_chat_administrator_custom_title,
-        ban_chat_sender_chat, unban_chat_sender_chat, set_chat_permissions,
-        export_chat_invite_link, create_chat_invite_link, edit_chat_invite_link,
-        revoke_chat_invite_link, set_chat_photo, delete_chat_photo, set_chat_title,
-        set_chat_description, pin_chat_message, unpin_chat_message, unpin_all_chat_messages,
-        leave_chat, get_chat, get_chat_administrators, get_chat_members_count, get_chat_member_count, get_chat_member,
-        set_chat_sticker_set, delete_chat_sticker_set, answer_callback_query,
-        set_my_commands, get_my_commands, delete_my_commands, answer_inline_query, edit_message_text,
-        edit_message_text_inline, edit_message_caption, edit_message_caption_inline,
-        edit_message_media, edit_message_media_inline, edit_message_reply_markup,
-        edit_message_reply_markup_inline, stop_poll, delete_message, send_sticker,
-        get_sticker_set, upload_sticker_file, create_new_sticker_set,
-        add_sticker_to_set, set_sticker_position_in_set, delete_sticker_from_set,
-        set_sticker_set_thumb, send_invoice, answer_shipping_query,
-        answer_pre_checkout_query, set_passport_data_errors, send_game,
-        set_game_score, set_game_score_inline, get_game_high_scores,
-        approve_chat_join_request, decline_chat_join_request => fwd_erased, fty
+        get_me,
+        log_out,
+        close,
+        get_updates,
+        set_webhook,
+        delete_webhook,
+        get_webhook_info,
+        forward_message,
+        copy_message,
+        send_message,
+        send_photo,
+        send_audio,
+        send_document,
+        send_video,
+        send_animation,
+        send_voice,
+        send_video_note,
+        send_media_group,
+        send_location,
+        edit_message_live_location,
+        edit_message_live_location_inline,
+        stop_message_live_location,
+        stop_message_live_location_inline,
+        send_venue,
+        send_contact,
+        send_poll,
+        send_dice,
+        send_chat_action,
+        get_user_profile_photos,
+        get_file,
+        kick_chat_member,
+        ban_chat_member,
+        unban_chat_member,
+        restrict_chat_member,
+        promote_chat_member,
+        set_chat_administrator_custom_title,
+        ban_chat_sender_chat,
+        unban_chat_sender_chat,
+        set_chat_permissions,
+        export_chat_invite_link,
+        create_chat_invite_link,
+        edit_chat_invite_link,
+        revoke_chat_invite_link,
+        set_chat_photo,
+        delete_chat_photo,
+        set_chat_title,
+        set_chat_description,
+        pin_chat_message,
+        unpin_chat_message,
+        unpin_all_chat_messages,
+        leave_chat,
+        get_chat,
+        get_chat_administrators,
+        get_chat_members_count,
+        get_chat_member_count,
+        get_chat_member,
+        set_chat_sticker_set,
+        delete_chat_sticker_set,
+        answer_callback_query,
+        set_my_commands,
+        get_my_commands,
+        set_chat_menu_button,
+        get_chat_menu_button,
+        set_my_default_administrator_rights,
+        get_my_default_administrator_rights,
+        delete_my_commands,
+        answer_inline_query,
+        answer_web_app_query,
+        edit_message_text,
+        edit_message_text_inline,
+        edit_message_caption,
+        edit_message_caption_inline,
+        edit_message_media,
+        edit_message_media_inline,
+        edit_message_reply_markup,
+        edit_message_reply_markup_inline,
+        stop_poll,
+        delete_message,
+        send_sticker,
+        get_sticker_set,
+        upload_sticker_file,
+        create_new_sticker_set,
+        add_sticker_to_set,
+        set_sticker_position_in_set,
+        delete_sticker_from_set,
+        set_sticker_set_thumb,
+        send_invoice,
+        answer_shipping_query,
+        answer_pre_checkout_query,
+        set_passport_data_errors,
+        send_game,
+        set_game_score,
+        set_game_score_inline,
+        get_game_high_scores,
+        approve_chat_join_request,
+        decline_chat_join_request
+        => fwd_erased, fty
     }
 }
 
@@ -523,6 +596,18 @@ trait ErasableRequester<'a> {
 
     fn get_my_commands(&self) -> ErasedRequest<'a, GetMyCommands, Self::Err>;
 
+    fn set_chat_menu_button(&self) -> ErasedRequest<'a, SetChatMenuButton, Self::Err>;
+
+    fn get_chat_menu_button(&self) -> ErasedRequest<'a, GetChatMenuButton, Self::Err>;
+
+    fn set_my_default_administrator_rights(
+        &self,
+    ) -> ErasedRequest<'a, SetMyDefaultAdministratorRights, Self::Err>;
+
+    fn get_my_default_administrator_rights(
+        &self,
+    ) -> ErasedRequest<'a, GetMyDefaultAdministratorRights, Self::Err>;
+
     fn delete_my_commands(&self) -> ErasedRequest<'a, DeleteMyCommands, Self::Err>;
 
     fn answer_inline_query(
@@ -530,6 +615,12 @@ trait ErasableRequester<'a> {
         inline_query_id: String,
         results: Vec<InlineQueryResult>,
     ) -> ErasedRequest<'a, AnswerInlineQuery, Self::Err>;
+
+    fn answer_web_app_query(
+        &self,
+        web_app_query_id: String,
+        result: InlineQueryResult,
+    ) -> ErasedRequest<'a, AnswerWebAppQuery, Self::Err>;
 
     fn edit_message_text(
         &self,
@@ -1169,6 +1260,26 @@ where
         Requester::get_my_commands(self).erase()
     }
 
+    fn set_chat_menu_button(&self) -> ErasedRequest<'a, SetChatMenuButton, Self::Err> {
+        Requester::set_chat_menu_button(self).erase()
+    }
+
+    fn get_chat_menu_button(&self) -> ErasedRequest<'a, GetChatMenuButton, Self::Err> {
+        Requester::get_chat_menu_button(self).erase()
+    }
+
+    fn set_my_default_administrator_rights(
+        &self,
+    ) -> ErasedRequest<'a, SetMyDefaultAdministratorRights, Self::Err> {
+        Requester::set_my_default_administrator_rights(self).erase()
+    }
+
+    fn get_my_default_administrator_rights(
+        &self,
+    ) -> ErasedRequest<'a, GetMyDefaultAdministratorRights, Self::Err> {
+        Requester::get_my_default_administrator_rights(self).erase()
+    }
+
     fn delete_my_commands(&self) -> ErasedRequest<'a, DeleteMyCommands, Self::Err> {
         Requester::delete_my_commands(self).erase()
     }
@@ -1179,6 +1290,14 @@ where
         results: Vec<InlineQueryResult>,
     ) -> ErasedRequest<'a, AnswerInlineQuery, Self::Err> {
         Requester::answer_inline_query(self, inline_query_id, results).erase()
+    }
+
+    fn answer_web_app_query(
+        &self,
+        web_app_query_id: String,
+        result: InlineQueryResult,
+    ) -> ErasedRequest<'a, AnswerWebAppQuery, Self::Err> {
+        Requester::answer_web_app_query(self, web_app_query_id, result).erase()
     }
 
     fn edit_message_text(