+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for Telegram Bot API [version 6.0](https://core.telegram.org/bots/api#april-16-2022)
   - Note that some field were renamed
 
+### Changed
+
+ - Accept `IntoIterator` in `KeyboardMarkup::append_row`.
+ - Accept `Into<String>` instead of `String` in `InlineKeyboardButton::{url, callback, switch_inline_query, switch_inline_query_current_chat}`.
+
 ## 0.5.1 - 2022-04-18
 
 ### Fixed