+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## 0.8.1 - 2022-04-24
+
 ### Added
 
  - Implement `GetChatId` for `Update`.